pyproject.toml
[tool.poetry]
name = "sergeichukd_gendiff"
version = "1.2.2"
description = ""
authors = ["Sergeichuk Dmitrii <sergeychukdmitry@gmail.com>"]
packages = [
	{include = "gendiff"},
    ]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"


[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.run:main"

[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/legacy/"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

